app/requirements.txt
flask
sqlalchemy
minio
pyjwt
bcrypt